rename exceptions to regions, continue iteration
