+++ gross_checks/ocean.vvhigh
@@ -5,7 +5,7 @@ taux       -3.0     3.0     -0.50    0.50
 tauy       -2.8     2.9     -0.50    0.50
 SSH        -2.80    1.60    -1.50    1.00
 SST        -1.96   36.0      1.00   30.0
-SSS        15.0    39.0     24.0    36.0
+SSS         5.0    39.0     22.0    36.0
 ePBL        0.75  2400.0   100.0   650.0
 MLD_003     0.50  3760.0    25.0  1750.0
 MLD_0125    0.60  4180.0    25.0  1750.0
